Pass correct reference expression for highlighting. Refactoring, add comments.
